+++ Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-authors = ["klaustinw"]
+authors = ["chcoomilk"]
 edition = "2018"
 name = "himitsu"
 version = "0.1.0"
@@ -9,13 +9,12 @@ version = "0.1.0"
 [dependencies]
 actix-cors = "0.5.4"
 actix-ratelimit = "0.3.1"
-actix-web = "3.3.2"
+actix-web = "3.3.3"
 argon2 = "0.3.2"
 diesel = {version = "1.4.8", features = ["postgres", "r2d2"]}
 dotenv = "0.15.0"
 env_logger = "0.9.0"
 magic-crypt = "3.1.9"
 r2d2 = "0.8.9"
-serde = "1.0.131"
-serde_json = "1.0.72"
-uuid = {version = "0.8.2", features = ["serde", "v4"]}
+serde = "1.0.133"
+serde_json = "1.0.74"
