+++ src/models.rs
@@ -1,2 +1,1 @@
 pub mod note;
-pub mod plain_note;
